pyproject.toml
[tool.poetry]
name = "course"
version = "0.1.0"
description = ""
authors = ["creedowl <creedowl@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.2"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
SQLAlchemy = "^1.4.18"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
uvicorn = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
